+++ src/main/java/com/se21/calbot/ClientManager/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This module contains all classes for user interacting entities like Discord, Telegram etc.
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.ClientManager;
