+++ src/main/java/com/se21/calbot/configs/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This module contains classes to manage application configurations
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.configs;
