+++ src/main/java/com/se21/calbot/controllers/AuthenticationController.java
@@ -6,6 +6,10 @@ import org.json.simple.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+/**
+ * This class serves as routing path controller. Any call to application/{route} will be handled here.
+ * For this project we need to expose one endpoint for Google Oauth 2.0 to share auth code. So that is being handled by this class.
+ */
 @RestController
 @CrossOrigin(origins = "*")
 public class AuthenticationController {
@@ -13,12 +17,13 @@ public class AuthenticationController {
     @Autowired
     CalendarFactory calendarFactory;
 
-    @RequestMapping(value = "/generate", method = RequestMethod.GET)
-    public String authenticateTest() {
-        Calendar calendar = calendarFactory.getCalendar("Google");
-        return calendar.authenticate("Xero978387");
-    }
-
+    /**
+     * This function will be executed whenever a call is made to https://{application endpoint}/test
+     * It is responsible to extract auth token and user id from url parameters and save it in the db for future use.
+     * @param code auth code
+     * @param state userId
+     * @return Msg for user
+     */
     @RequestMapping(value = "/test", method = RequestMethod.GET)
     public String token(@RequestParam String code, @RequestParam String state) {
         Calendar calendar = calendarFactory.getCalendar("Google");
@@ -26,6 +31,20 @@ public class AuthenticationController {
         return "Auth completed successfully, please close this window and get back to discord bot!";
     }
 
+    /**
+     * Test endpoints, feel free to use them during development
+     * @throws Exception
+     */
+    @RequestMapping(value = "/generate", method = RequestMethod.GET)
+    public String authenticateTest() {
+        Calendar calendar = calendarFactory.getCalendar("Google");
+        return calendar.authenticate("Xero978387");
+    }
+
+    /**
+     * Test endpoints, feel free to use them during development
+     * @throws Exception
+     */
     @RequestMapping(value = "/retrieve", method = RequestMethod.GET)
     public void summary() throws Exception{
         Calendar calendar = calendarFactory.getCalendar("Google");
