+++ src/main/java/com/se21/calbot/controllers/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This module contains classes to control various entities(aPAS, auth, testPing)
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.controllers;
