+++ src/main/java/com/se21/calbot/enums/Enums.java
@@ -1,7 +1,13 @@
 package com.se21.calbot.enums;
 
+/**
+ * Enum manager class. It contains all enums needed in the project at one place
+ */
 public class Enums {
 
+    /**
+     * Response send by calendar classes
+     */
     public enum calApiResponse{
         Success,
         Failure,
@@ -11,12 +17,18 @@ public class Enums {
         //Keep adding more to enrich debugging arsenal
     }
 
+    /**
+     * Type of calendar being used by any user can be represented using this enum
+     */
     public enum calendarType{
         googleCalendar,
         appleCalendar,
         samsungCalendar,
     }
 
+    /**
+     * All available operations types.
+     */
     public enum operationType{
         Add,
         Update,
