+++ src/main/java/com/se21/calbot/enums/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains enum classes used in project
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.enums;
