+++ src/main/java/com/se21/calbot/factories/CalendarFactory.java
@@ -6,12 +6,21 @@ import com.se21.calbot.services.GoogleCalendarService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+/**
+ * Factory class to generate object for type of calendar used by user.
+ * This class is one of responsible to help achieve factory pattern in project.
+ */
 @Component
 public class CalendarFactory {
 
     @Autowired
     GoogleCalendarService googleCalendarService;
 
+    /**
+     * Provides object of calendar used by current user
+     * @param channel requested calendar type
+     * @return specific calendar type object will be typecast with Calendar interface class (like GoogleCalendarService obj will be typecast to Calendar object)
+     */
     public Calendar getCalendar(String channel) {
 
         switch(channel) {
