+++ src/main/java/com/se21/calbot/factories/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This module contains factory classes, and it helps in achieving factory
+ * pattern, which in turn helps follow the design principles.
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.factories;
