+++ src/main/java/com/se21/calbot/interfaces/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains all the interfaces
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.interfaces;
