+++ src/main/java/com/se21/calbot/listeners/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This module contains message listener classes to listen from user interacting entities
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.listeners;
