+++ src/main/java/com/se21/calbot/model/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains DB schema
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.model;
