+++ src/main/java/com/se21/calbot/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * This package is the highest package in package hierarchy, containing all of other packages and whole code.
+ * <p>
+ * All further subpackages lie in this package.
+ * </p>
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot;
