+++ src/main/java/com/se21/calbot/repositories/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Db manager classes
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.repositories;
