+++ src/main/java/com/se21/calbot/services/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Service classes to implement functionalities
+ *
+ * @since 1.0
+ * @version 1.0
+ */
+package com.se21.calbot.services;
